requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
APScheduler==3.11.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
av==15.1.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
bleach==6.2.0
blis==1.3.0
build==1.3.0
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chromadb==1.2.1
click==8.3.0
cloudpathlib==0.23.0
coloredlogs==15.0.1
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
durationpy==0.10
einops==0.8.1
elevenlabs==2.17.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.1
faiss-cpu==1.12.0
fastjsonschema==2.21.2
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
google-auth==2.41.1
googleapis-common-protos==1.71.0
grpcio==1.76.0
h11==0.14.0
hf-xet==1.1.9
httpcore==0.17.3
httptools==0.7.1
httpx==0.24.1
httpx-sse==0.4.3
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imagebind @ file:///Users/mbarkalywsf/PycharmProjects/Agentic/ImageBind-main
importlib_metadata==8.7.0
importlib_resources==6.5.2
iopath==0.1.10
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==0.3.27
langchain-chroma==1.0.0
langchain-community==0.3.31
langchain-core==1.0.2
langchain-experimental==0.3.4
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langcodes==3.5.0
langsmith==0.4.31
language_data==1.3.0
lark==1.3.0
llama_cpp_python==0.3.16
llvmlite==0.44.0
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.13
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.3.0
openai-whisper==20250625
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.2.250926
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.5
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
portalocker==3.2.0
posthog==5.4.0
preshed==3.0.10
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-trans==0.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-pptx==1.0.2
python-telegram-bot==20.3
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@6cdc929315aab1b5674b6dcf73b16ec99147735f
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.1
Send2Trash==1.8.3
sentence-transformers==5.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
terminado==0.18.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.11.0
timm==1.0.20
tinycss2==1.4.0
tokenizers==0.22.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
typer==0.19.2
types-python-dateutil==2.9.0.20251008
types-pytz==2025.2.0.20250809
types-regex==2025.9.18.20250921
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.38.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.14
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wikipedia==1.4.0
wrapt==1.17.3
xlsxwriter==3.2.9
xx_ent_wiki_sm @ https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.8.0/xx_ent_wiki_sm-3.8.0-py3-none-any.whl#sha256=6f3c4b853852ea9e9d2dc76cc950dddb10a7e4c42d813308caefe6c5e8be2f0a
yacs==0.1.8
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
